requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6


# Django
django==2.2


# Environment
django-environ==0.4.5

# JWT
pyjwt==1.7.1

# Static files
whitenoise==4.1.2

# Celery
redis>=2.10.6, < 3
django-redis==4.10.0
celery==4.2.1
flower==0.9.2

# DJANGO REST FRAMEWORK---------------------------------------------------
djangorestframework==3.12.4

swapper==1.1.2